requirements-dev.txt
# Testing Framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0          # Parallel testing
pytest-timeout>=2.1.0         # Test timeout
pytest-benchmark>=4.0.0       # Performance benchmarking
pytest-mock>=3.11.0           # Mock fixtures
pytest-html>=3.2.0            # HTML reports
pytest-json-report>=1.5.0     # JSON reports
pytest-monitor>=1.6.6         # Test performance tracking

# Property-Based Testing
hypothesis>=6.82.0
hypothesis[numpy]>=6.82.0
hypothesis[pandas]>=6.82.0

# Mocking & Fixtures
faker>=19.3.0                 # Fake data generation
factory-boy>=3.3.0            # Factory fixtures
freezegun>=1.2.2              # Time mocking
responses>=0.23.0             # HTTP mocking

# Contract & Schema Validation
jsonschema>=4.19.0            # JSON Schema validation
avro-python3>=1.10.0          # Avro schema support
pydantic>=2.3.0               # Data validation

# Code Quality & Linting
black>=23.7.0                 # Code formatter
flake8>=6.1.0                 # Linter
mypy>=1.5.0                   # Type checker
isort>=5.12.0                 # Import sorter
pylint>=2.17.0                # Static analyzer
bandit>=1.7.5                 # Security linting
safety>=2.3.0                 # Dependency vulnerability scanner
pip-audit>=2.6.0              # Dependency auditing

# Coverage Tools
coverage[toml]>=7.3.0         # Coverage with pyproject.toml support

# Performance & Profiling
memory-profiler>=0.61.0       # Memory profiling
py-spy>=0.3.14                # Sampling profiler
locust>=2.15.0                # Load testing (optional)

# Documentation
sphinx>=7.1.0                 # Documentation generator
sphinx-rtd-theme>=1.3.0       # ReadTheDocs theme
sphinx-autodoc-typehints>=1.24.0

# CI/CD Tools
tox>=4.6.0                    # Test automation
pre-commit>=3.3.0             # Git hooks

# Chaos Engineering (custom - no standard lib)
# Will implement custom chaos testing framework

# Monitoring & Observability (for testing)
prometheus-client>=0.17.0     # Metrics export

# Utilities
pyyaml>=6.0                   # YAML parsing
python-dotenv>=1.0.0          # Environment variables
tabulate>=0.9.0               # Table formatting
colorama>=0.4.6               # Colored terminal output
